+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 next
------
-* Support `Hashable` from `hashable >= 1.2.5.0`
+----
+* Replace use of `Hashable1` from `hashable-extras` in favor of `Hashable` from
+  `hashable-1.2.5.0`. As a result, the `hashable-extras` dependency has been removed.
 * Revamp `Setup.hs` to use `cabal-doctest`. This makes it build
   with `Cabal-2.0`, and makes the `doctest`s work with `cabal new-build` and
   sandboxes.